refactor: simplify code
I was having trouble getting AsRef to work the way I thought it should
with the paths. I finally settled on the fact that the code should take
only what it needs, and that means that at its core, it needs Path
structs. Refactoring to take that made things simpler and clearer in the
code and when calling it.

I also moved the tests to the top of the lib, placing them front and
center, giving them a higher perceived priority over the code itself,
which aligns better with what I want to communicate in the code.